auth/djangoRest/onePageWebAppTokenHeader/requirements.txt
Django==3.1.6
djangorestframework

